third_party/blink/web_tests/external/wpt/css/css-values/calc-infinity-nan-computed-expected.txt

This is a testharness.js-based test.
FAIL Property width value 'calc(NaN * 1px)' assert_equals: expected "0px" but got "3.35544e+07px"
PASS Property width value 'calc(infinity * 1px)'
PASS Property width value 'calc(infinity * 1cm)'
FAIL Property width value 'calc(NaN * 1rem)' assert_equals: expected "0px" but got "3.35544e+07px"
FAIL Property width value 'calc(10.135262721212548pc - 199pt / NaN)' assert_equals: expected "0px" but got "3.35544e+07px"
FAIL Property width value 'max(15px, NaN * 1px)' assert_equals: expected "0px" but got "3.35544e+07px"
FAIL Property width value 'max(NaN * 1px, 15px)' assert_equals: expected "0px" but got "3.35544e+07px"
FAIL Property width value 'min(15px, NaN * 1px)' assert_equals: expected "0px" but got "3.35544e+07px"
FAIL Property width value 'min(NaN * 1px, 15px)' assert_equals: expected "0px" but got "3.35544e+07px"
FAIL Property width value 'calc(infinity * 1px - infinity * 1%)' assert_equals: expected "0px" but got "3.35544e+07px"
FAIL Property width value 'calc(infinity * 1px + infinity * 1%)' assert_equals: expected "0px" but got "3.35544e+07px"
FAIL Property width value 'calc(min(NaN * 1px, infinity * 1px) + max(infinity * 1px, -infinity * 1px))' assert_equals: expected "0px" but got "3.35544e+07px"
FAIL Property width value 'calc(infinity * 1px - max(infinity * 1%, 0%))' assert_equals: expected "0px" but got "3.35544e+07px"
PASS Property width value 'calc(max(infinity * 1px, 10px))'
PASS Property margin-left value 'calc(-infinity * 1px)'
PASS Property margin-left value 'calc(min(1px, -infinity * 1%))'
PASS Property margin-left value 'calc(-infinity * 1%)'
PASS Property margin-left value 'calc(max(10000px, 0px) + min(-infinity * 1px, infinity * 1px))'
PASS Property margin-left value 'calc(-infinity * 1px - infinity * 1px)'
PASS Property margin-left value 'calc(min(-infinity * 1px, 10px))'
FAIL Property animation-duration value 'calc(NaN * 1s)' assert_equals: expected "0s" but got "1.79769e+308s"
PASS Property animation-duration value 'calc(infinity * 1s)'
FAIL Property animation-duration value 'calc(1 / 0 * 1s)' assert_equals: expected "0s" but got "1.79769e+308s"
PASS Property animation-duration value 'calc(max(infinity * 1s, 10s)'
PASS Property transition-delay value 'calc(-infinity* 1s)'
PASS Property transition-delay value 'calc(max(10000s, 0s) + min(-infinity * 1s, infinity * 1s))'
PASS Property transition-delay value 'calc(min(-infinity * 1s, 10s))'
PASS Property rotate(calc(infinity * 1deg)) value expected same with rotate(0deg) in +/-0.0001
PASS Property rotate(calc(-infinity * 1deg)) value expected same with rotate(0deg) in +/-0.0001
PASS Property rotate(calc(NaN * 1deg)) value expected same with rotate(0deg) in +/-0.0001
PASS Property rotate(calc(infinity * 1turn)) value expected same with rotate(0deg) in +/-0.0001
PASS Property rotate(calc(-infinity * 1turn)) value expected same with rotate(0deg) in +/-0.0001
PASS Property rotate(calc(NaN * 1turn)) value expected same with rotate(0deg) in +/-0.0001
PASS Property rotate(calc(infinity * 1rad)) value expected same with rotate(0deg) in +/-0.0001
PASS Property rotate(calc(-infinity * 1rad)) value expected same with rotate(0deg) in +/-0.0001
PASS Property rotate(calc(NaN * 1rad)) value expected same with rotate(0deg) in +/-0.0001
PASS Property rotate(calc(infinity * 1grad)) value expected same with rotate(0deg) in +/-0.0001
PASS Property rotate(calc(-infinity * 1grad)) value expected same with rotate(0deg) in +/-0.0001
PASS Property rotate(calc(NaN * 1grad)) value expected same with rotate(0deg) in +/-0.0001
Harness: the test ran to completion.
